+++ x/wasm/dynamic_link_test.go
@@ -1,13 +1,10 @@
-package wasmplus
+package wasm
 
 import (
 	"fmt"
 	"testing"
 
-	sdk "github.com/line/lbm-sdk/types"
-	"github.com/line/wasmd/x/wasm"
 	"github.com/line/wasmd/x/wasm/keeper"
-	"github.com/line/wasmd/x/wasmplus/types"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	abci "github.com/tendermint/tendermint/abci/types"
@@ -30,7 +27,7 @@ func TestDynamicPingPongWorks(t *testing.T) {
 	h := data.module.Route().Handler()
 
 	// store dynamic callee code
-	storeCalleeMsg := &wasm.MsgStoreCode{
+	storeCalleeMsg := &MsgStoreCode{
 		Sender:       addr1,
 		WASMByteCode: calleeContract,
 	}
@@ -41,7 +38,7 @@ func TestDynamicPingPongWorks(t *testing.T) {
 	assertStoreCodeResponse(t, res.Data, calleeCodeId)
 
 	// store dynamic caller code
-	storeCallerMsg := &wasm.MsgStoreCode{
+	storeCallerMsg := &MsgStoreCode{
 		Sender:       addr1,
 		WASMByteCode: callerContract,
 	}
@@ -52,7 +49,7 @@ func TestDynamicPingPongWorks(t *testing.T) {
 	assertStoreCodeResponse(t, res.Data, callerCodeId)
 
 	// instantiate callee contract
-	instantiateCalleeMsg := &wasm.MsgInstantiateContract{
+	instantiateCalleeMsg := &MsgInstantiateContract{
 		Sender: addr1,
 		CodeID: calleeCodeId,
 		Label:  "callee",
@@ -66,7 +63,7 @@ func TestDynamicPingPongWorks(t *testing.T) {
 
 	// instantiate caller contract
 	cosmwasmInstantiateCallerMsg := fmt.Sprintf(`{"callee_addr":"%s"}`, calleeContractAddress)
-	instantiateCallerMsg := &wasm.MsgInstantiateContract{
+	instantiateCallerMsg := &MsgInstantiateContract{
 		Sender: addr1,
 		CodeID: callerCodeId,
 		Label:  "caller",
@@ -80,7 +77,7 @@ func TestDynamicPingPongWorks(t *testing.T) {
 
 	// execute ping
 	cosmwasmExecuteMsg := `{"ping":{"ping_num":"100"}}`
-	executeMsg := wasm.MsgExecuteContract{
+	executeMsg := MsgExecuteContract{
 		Sender:   addr1,
 		Contract: callerContractAddress,
 		Msg:      []byte(cosmwasmExecuteMsg),
@@ -107,7 +104,7 @@ func TestDynamicReEntrancyFails(t *testing.T) {
 	h := data.module.Route().Handler()
 
 	// store dynamic callee code
-	storeCalleeMsg := &wasm.MsgStoreCode{
+	storeCalleeMsg := &MsgStoreCode{
 		Sender:       addr1,
 		WASMByteCode: calleeContract,
 	}
@@ -118,7 +115,7 @@ func TestDynamicReEntrancyFails(t *testing.T) {
 	assertStoreCodeResponse(t, res.Data, calleeCodeId)
 
 	// store dynamic caller code
-	storeCallerMsg := &wasm.MsgStoreCode{
+	storeCallerMsg := &MsgStoreCode{
 		Sender:       addr1,
 		WASMByteCode: callerContract,
 	}
@@ -129,7 +126,7 @@ func TestDynamicReEntrancyFails(t *testing.T) {
 	assertStoreCodeResponse(t, res.Data, callerCodeId)
 
 	// instantiate callee contract
-	instantiateCalleeMsg := &wasm.MsgInstantiateContract{
+	instantiateCalleeMsg := &MsgInstantiateContract{
 		Sender: addr1,
 		CodeID: calleeCodeId,
 		Label:  "callee",
@@ -143,7 +140,7 @@ func TestDynamicReEntrancyFails(t *testing.T) {
 
 	// instantiate caller contract
 	cosmwasmInstantiateCallerMsg := fmt.Sprintf(`{"callee_addr":"%s"}`, calleeContractAddress)
-	instantiateCallerMsg := &wasm.MsgInstantiateContract{
+	instantiateCallerMsg := &MsgInstantiateContract{
 		Sender: addr1,
 		CodeID: callerCodeId,
 		Label:  "caller",
@@ -157,7 +154,7 @@ func TestDynamicReEntrancyFails(t *testing.T) {
 
 	// execute ping
 	cosmwasmExecuteMsg := `{"try_re_entrancy":{}}`
-	executeMsg := wasm.MsgExecuteContract{
+	executeMsg := MsgExecuteContract{
 		Sender:   addr1,
 		Contract: callerContractAddress,
 		Msg:      []byte(cosmwasmExecuteMsg),
@@ -174,7 +171,7 @@ func TestDynamicLinkInterfaceValidation(t *testing.T) {
 	h := data.module.Route().Handler()
 
 	// store dynamic callee code
-	storeCalleeMsg := &wasm.MsgStoreCode{
+	storeCalleeMsg := &MsgStoreCode{
 		Sender:       addr1,
 		WASMByteCode: calleeContract,
 	}
@@ -185,7 +182,7 @@ func TestDynamicLinkInterfaceValidation(t *testing.T) {
 	assertStoreCodeResponse(t, res.Data, calleeCodeId)
 
 	// store dynamic caller code
-	storeCallerMsg := &wasm.MsgStoreCode{
+	storeCallerMsg := &MsgStoreCode{
 		Sender:       addr1,
 		WASMByteCode: callerContract,
 	}
@@ -196,7 +193,7 @@ func TestDynamicLinkInterfaceValidation(t *testing.T) {
 	assertStoreCodeResponse(t, res.Data, callerCodeId)
 
 	// instantiate callee contract
-	instantiateCalleeMsg := &wasm.MsgInstantiateContract{
+	instantiateCalleeMsg := &MsgInstantiateContract{
 		Sender: addr1,
 		CodeID: calleeCodeId,
 		Label:  "callee",
@@ -210,7 +207,7 @@ func TestDynamicLinkInterfaceValidation(t *testing.T) {
 
 	// instantiate caller contract
 	cosmwasmInstantiateCallerMsg := fmt.Sprintf(`{"callee_addr":"%s"}`, calleeContractAddress)
-	instantiateCallerMsg := &wasm.MsgInstantiateContract{
+	instantiateCallerMsg := &MsgInstantiateContract{
 		Sender: addr1,
 		CodeID: callerCodeId,
 		Label:  "caller",
@@ -224,7 +221,7 @@ func TestDynamicLinkInterfaceValidation(t *testing.T) {
 
 	// execute validate interface
 	cosmwasmExecuteMsg := `{"validate_interface":{}}`
-	executeMsg := wasm.MsgExecuteContract{
+	executeMsg := MsgExecuteContract{
 		Sender:   addr1,
 		Contract: callerContractAddress,
 		Msg:      []byte(cosmwasmExecuteMsg),
@@ -235,7 +232,7 @@ func TestDynamicLinkInterfaceValidation(t *testing.T) {
 
 	// execute validate interface error
 	cosmwasmExecuteMsgErr := `{"validate_interface_err":{}}`
-	executeMsgErr := wasm.MsgExecuteContract{
+	executeMsgErr := MsgExecuteContract{
 		Sender:   addr1,
 		Contract: callerContractAddress,
 		Msg:      []byte(cosmwasmExecuteMsgErr),
@@ -255,7 +252,7 @@ func TestDynamicCallAndTraditionalQueryWork(t *testing.T) {
 	q := data.module.LegacyQuerierHandler(nil)
 
 	// store callee code (number)
-	storeCalleeMsg := &wasm.MsgStoreCode{
+	storeCalleeMsg := &MsgStoreCode{
 		Sender:       addr1,
 		WASMByteCode: numberContract,
 	}
@@ -266,7 +263,7 @@ func TestDynamicCallAndTraditionalQueryWork(t *testing.T) {
 	assertStoreCodeResponse(t, res.Data, calleeCodeId)
 
 	// store caller code (call-number)
-	storeCallerMsg := &wasm.MsgStoreCode{
+	storeCallerMsg := &MsgStoreCode{
 		Sender:       addr1,
 		WASMByteCode: callNumberContract,
 	}
@@ -277,7 +274,7 @@ func TestDynamicCallAndTraditionalQueryWork(t *testing.T) {
 	assertStoreCodeResponse(t, res.Data, callerCodeId)
 
 	// instantiate callee contract
-	instantiateCalleeMsg := &wasm.MsgInstantiateContract{
+	instantiateCalleeMsg := &MsgInstantiateContract{
 		Sender: addr1,
 		CodeID: calleeCodeId,
 		Label:  "number",
@@ -291,7 +288,7 @@ func TestDynamicCallAndTraditionalQueryWork(t *testing.T) {
 
 	// instantiate caller contract
 	cosmwasmInstantiateCallerMsg := fmt.Sprintf(`{"callee_addr":"%s"}`, calleeContractAddress)
-	instantiateCallerMsg := &wasm.MsgInstantiateContract{
+	instantiateCallerMsg := &MsgInstantiateContract{
 		Sender: addr1,
 		CodeID: callerCodeId,
 		Label:  "call-number",
@@ -305,7 +302,7 @@ func TestDynamicCallAndTraditionalQueryWork(t *testing.T) {
 
 	// traditional queries from caller
 	queryPath := []string{
-		wasm.QueryGetContractState,
+		QueryGetContractState,
 		callerContractAddress,
 		keeper.QueryMethodContractStateSmart,
 	}
@@ -322,7 +319,7 @@ func TestDynamicCallAndTraditionalQueryWork(t *testing.T) {
 
 	// execute mul
 	cosmwasmExecuteMsg := `{"mul":{"value":2}}`
-	executeMsg := wasm.MsgExecuteContract{
+	executeMsg := MsgExecuteContract{
 		Sender:   addr1,
 		Contract: callerContractAddress,
 		Msg:      []byte(cosmwasmExecuteMsg),
@@ -355,7 +352,7 @@ func TestDynamicCallWithWriteFailsByQuery(t *testing.T) {
 	q := data.module.LegacyQuerierHandler(nil)
 
 	// store callee code (number)
-	storeCalleeMsg := &wasm.MsgStoreCode{
+	storeCalleeMsg := &MsgStoreCode{
 		Sender:       addr1,
 		WASMByteCode: numberContract,
 	}
@@ -366,7 +363,7 @@ func TestDynamicCallWithWriteFailsByQuery(t *testing.T) {
 	assertStoreCodeResponse(t, res.Data, calleeCodeId)
 
 	// store caller code (call-number)
-	storeCallerMsg := &wasm.MsgStoreCode{
+	storeCallerMsg := &MsgStoreCode{
 		Sender:       addr1,
 		WASMByteCode: callNumberContract,
 	}
@@ -377,7 +374,7 @@ func TestDynamicCallWithWriteFailsByQuery(t *testing.T) {
 	assertStoreCodeResponse(t, res.Data, callerCodeId)
 
 	// instantiate callee contract
-	instantiateCalleeMsg := &wasm.MsgInstantiateContract{
+	instantiateCalleeMsg := &MsgInstantiateContract{
 		Sender: addr1,
 		CodeID: calleeCodeId,
 		Label:  "number",
@@ -391,7 +388,7 @@ func TestDynamicCallWithWriteFailsByQuery(t *testing.T) {
 
 	// instantiate caller contract
 	cosmwasmInstantiateCallerMsg := fmt.Sprintf(`{"callee_addr":"%s"}`, calleeContractAddress)
-	instantiateCallerMsg := &wasm.MsgInstantiateContract{
+	instantiateCallerMsg := &MsgInstantiateContract{
 		Sender: addr1,
 		CodeID: callerCodeId,
 		Label:  "call-number",
@@ -405,7 +402,7 @@ func TestDynamicCallWithWriteFailsByQuery(t *testing.T) {
 
 	// query which tries to write value to storage
 	queryPath := []string{
-		wasm.QueryGetContractState,
+		QueryGetContractState,
 		callerContractAddress,
 		keeper.QueryMethodContractStateSmart,
 	}
@@ -422,7 +419,7 @@ func TestDynamicCallCalleeFails(t *testing.T) {
 	h := data.module.Route().Handler()
 
 	// store dynamic callee code
-	storeCalleeMsg := &wasm.MsgStoreCode{
+	storeCalleeMsg := &MsgStoreCode{
 		Sender:       addr1,
 		WASMByteCode: calleeContract,
 	}
@@ -433,7 +430,7 @@ func TestDynamicCallCalleeFails(t *testing.T) {
 	assertStoreCodeResponse(t, res.Data, calleeCodeId)
 
 	// store dynamic caller code
-	storeCallerMsg := &wasm.MsgStoreCode{
+	storeCallerMsg := &MsgStoreCode{
 		Sender:       addr1,
 		WASMByteCode: callerContract,
 	}
@@ -444,7 +441,7 @@ func TestDynamicCallCalleeFails(t *testing.T) {
 	assertStoreCodeResponse(t, res.Data, callerCodeId)
 
 	// instantiate callee contract
-	instantiateCalleeMsg := &wasm.MsgInstantiateContract{
+	instantiateCalleeMsg := &MsgInstantiateContract{
 		Sender: addr1,
 		CodeID: calleeCodeId,
 		Label:  "callee",
@@ -458,7 +455,7 @@ func TestDynamicCallCalleeFails(t *testing.T) {
 
 	// instantiate caller contract
 	cosmwasmInstantiateCallerMsg := fmt.Sprintf(`{"callee_addr":"%s"}`, calleeContractAddress)
-	instantiateCallerMsg := &wasm.MsgInstantiateContract{
+	instantiateCallerMsg := &MsgInstantiateContract{
 		Sender: addr1,
 		CodeID: callerCodeId,
 		Label:  "caller",
@@ -472,7 +469,7 @@ func TestDynamicCallCalleeFails(t *testing.T) {
 
 	// execute do_panic
 	cosmwasmExecuteMsg := `{"do_panic":{}}`
-	executeMsg := wasm.MsgExecuteContract{
+	executeMsg := MsgExecuteContract{
 		Sender:   addr1,
 		Contract: callerContractAddress,
 		Msg:      []byte(cosmwasmExecuteMsg),
@@ -481,172 +478,4 @@ func TestDynamicCallCalleeFails(t *testing.T) {
 	res, err = h(data.ctx, &executeMsg)
 	assert.ErrorContains(t, err, "Error in dynamic link")
 	assert.ErrorContains(t, err, "RuntimeError: unreachable")
-}
-
-// This tests callCallablePoint in inactive contract
-func TestCallInActiveContractFails(t *testing.T) {
-	// setup
-	data := setupTest(t)
-
-	h := data.module.Route().Handler()
-
-	// store dynamic callee code
-	storeCalleeMsg := &wasm.MsgStoreCode{
-		Sender:       addr1,
-		WASMByteCode: calleeContract,
-	}
-	res, err := h(data.ctx, storeCalleeMsg)
-	require.NoError(t, err)
-
-	calleeCodeId := uint64(1)
-	assertStoreCodeResponse(t, res.Data, calleeCodeId)
-
-	// store dynamic caller code
-	storeCallerMsg := &wasm.MsgStoreCode{
-		Sender:       addr1,
-		WASMByteCode: callerContract,
-	}
-	res, err = h(data.ctx, storeCallerMsg)
-	require.NoError(t, err)
-
-	callerCodeId := uint64(2)
-	assertStoreCodeResponse(t, res.Data, callerCodeId)
-
-	// instantiate callee contract
-	instantiateCalleeMsg := &wasm.MsgInstantiateContract{
-		Sender: addr1,
-		CodeID: calleeCodeId,
-		Label:  "callee",
-		Msg:    []byte(`{}`),
-		Funds:  nil,
-	}
-	res, err = h(data.ctx, instantiateCalleeMsg)
-	require.NoError(t, err)
-
-	calleeContractAddress := parseInitResponse(t, res.Data)
-
-	// instantiate caller contract
-	cosmwasmInstantiateCallerMsg := fmt.Sprintf(`{"callee_addr":"%s"}`, calleeContractAddress)
-	instantiateCallerMsg := &wasm.MsgInstantiateContract{
-		Sender: addr1,
-		CodeID: callerCodeId,
-		Label:  "caller",
-		Msg:    []byte(cosmwasmInstantiateCallerMsg),
-		Funds:  nil,
-	}
-	res, err = h(data.ctx, instantiateCallerMsg)
-	require.NoError(t, err)
-
-	callerContractAddress := parseInitResponse(t, res.Data)
-
-	// add contract to inactive
-	src := types.DeactivateContractProposal{
-		Title:       "Foo",
-		Description: "Bar",
-		Contract:    calleeContractAddress,
-	}
-	storedProposal, err := data.govKeeper.SubmitProposal(data.ctx, &src)
-	require.NoError(t, err)
-	handler := data.govKeeper.Router().GetRoute(storedProposal.ProposalRoute())
-	err = handler(data.ctx.WithEventManager(sdk.NewEventManager()), storedProposal.GetContent())
-	require.NoError(t, err)
-	calleeContractSdkAddress, err := sdk.AccAddressFromBech32(calleeContractAddress)
-	require.NoError(t, err)
-	isInactive := data.keeper.IsInactiveContract(data.ctx, calleeContractSdkAddress)
-	require.True(t, isInactive)
-
-	// execute ping
-	cosmwasmExecuteMsg := `{"ping":{"ping_num":"100"}}`
-	executeMsg := wasm.MsgExecuteContract{
-		Sender:   addr1,
-		Contract: callerContractAddress,
-		Msg:      []byte(cosmwasmExecuteMsg),
-		Funds:    nil,
-	}
-	_, err = h(data.ctx, &executeMsg)
-	assert.ErrorContains(t, err, "called contract cannot be executed")
-}
-
-// This tests validateInterface in inactive contract
-func TestValidateInActiveContractFails(t *testing.T) {
-	// setup
-	data := setupTest(t)
-
-	h := data.module.Route().Handler()
-
-	// store dynamic callee code
-	storeCalleeMsg := &wasm.MsgStoreCode{
-		Sender:       addr1,
-		WASMByteCode: calleeContract,
-	}
-	res, err := h(data.ctx, storeCalleeMsg)
-	require.NoError(t, err)
-
-	calleeCodeId := uint64(1)
-	assertStoreCodeResponse(t, res.Data, calleeCodeId)
-
-	// store dynamic caller code
-	storeCallerMsg := &wasm.MsgStoreCode{
-		Sender:       addr1,
-		WASMByteCode: callerContract,
-	}
-	res, err = h(data.ctx, storeCallerMsg)
-	require.NoError(t, err)
-
-	callerCodeId := uint64(2)
-	assertStoreCodeResponse(t, res.Data, callerCodeId)
-
-	// instantiate callee contract
-	instantiateCalleeMsg := &wasm.MsgInstantiateContract{
-		Sender: addr1,
-		CodeID: calleeCodeId,
-		Label:  "callee",
-		Msg:    []byte(`{}`),
-		Funds:  nil,
-	}
-	res, err = h(data.ctx, instantiateCalleeMsg)
-	require.NoError(t, err)
-
-	calleeContractAddress := parseInitResponse(t, res.Data)
-
-	// instantiate caller contract
-	cosmwasmInstantiateCallerMsg := fmt.Sprintf(`{"callee_addr":"%s"}`, calleeContractAddress)
-	instantiateCallerMsg := &wasm.MsgInstantiateContract{
-		Sender: addr1,
-		CodeID: callerCodeId,
-		Label:  "caller",
-		Msg:    []byte(cosmwasmInstantiateCallerMsg),
-		Funds:  nil,
-	}
-	res, err = h(data.ctx, instantiateCallerMsg)
-	require.NoError(t, err)
-
-	callerContractAddress := parseInitResponse(t, res.Data)
-
-	// add contract to inactive
-	src := types.DeactivateContractProposal{
-		Title:       "Foo",
-		Description: "Bar",
-		Contract:    calleeContractAddress,
-	}
-	storedProposal, err := data.govKeeper.SubmitProposal(data.ctx, &src)
-	require.NoError(t, err)
-	handler := data.govKeeper.Router().GetRoute(storedProposal.ProposalRoute())
-	err = handler(data.ctx.WithEventManager(sdk.NewEventManager()), storedProposal.GetContent())
-	require.NoError(t, err)
-	calleeContractSdkAddress, err := sdk.AccAddressFromBech32(calleeContractAddress)
-	require.NoError(t, err)
-	isInactive := data.keeper.IsInactiveContract(data.ctx, calleeContractSdkAddress)
-	require.True(t, isInactive)
-
-	// execute validate interface
-	cosmwasmExecuteMsg := `{"validate_interface":{}}`
-	executeMsg := wasm.MsgExecuteContract{
-		Sender:   addr1,
-		Contract: callerContractAddress,
-		Msg:      []byte(cosmwasmExecuteMsg),
-		Funds:    nil,
-	}
-	_, err = h(data.ctx, &executeMsg)
-	assert.ErrorContains(t, err, "try to validate a contract cannot be executed")
 }