+++ x/wasm/types/events.go
@@ -5,6 +5,10 @@ const (
 	WasmModuleEventType = "wasm"
 	// CustomContractEventPrefix contracts can create custom events. To not mix them with other system events they got the `wasm-` prefix.
 	CustomContractEventPrefix = "wasm-"
+	// events from callable point
+	CallablePointEventType = "wasm-callablepoint"
+	// prefix for custom events from callable point
+	CustomCallablePointEventPrefix = "wasm-callablepoint-"
 
 	EventTypeStoreCode         = "store_code"
 	EventTypeInstantiate       = "instantiate"
@@ -21,9 +25,12 @@ const (
 const (
 	AttributeReservedPrefix = "_"
 
+	AttributeKeyCallstack          = "_callstack"
 	AttributeKeyContractAddr       = "_contract_address"
 	AttributeKeyCodeID             = "code_id"
+	AttributeKeyCodeIDs            = "code_ids"
 	AttributeKeyChecksum           = "code_checksum"
+	AttributeKeyFeature            = "feature"
 	AttributeKeyResultDataHex      = "result"
 	AttributeKeyRequiredCapability = "required_capability"
 )